+++ generators/app/src/end.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var chalk = require('chalk');
+const chalk = require('chalk');
 
 module.exports = function() {
-  this.log(chalk.green('\n Your plugin has been successfully installed ! \n'));
+  this.log(chalk.green('\n Your plugin has been successfully installed! \n'));
 };