+++ generators/app/src/prompting.js
@@ -1,14 +1,15 @@
 'use strict';
 
-var chalk = require('chalk');
-var yosay = require('yosay');
-var _ = require('lodash');
-var paths = require('../config/paths');
+const chalk = require('chalk');
+const yosay = require('yosay');
+const _ = require('lodash');
+
+const paths = require('../config/paths');
 
 /**
  * Validates the given plugin name.
  *
- *  Plugin name should :
+ *  Plugin name should:
  *  - Not start or end with a dash
  *  - Start by openveo
  *  - Contains other characters than latin letters and dashes
@@ -20,14 +21,13 @@ function isValidPluginName(name) {
   return /^(?!openveo|-)[A-Za-z-]+[^-]$/.test(name);
 }
 
-module.exports = function() {
-  var self = this;
-  var prompts = [];
+module.exports = async function() {
+  const prompts = [];
 
   // Have Yeoman greet the user.
   this.log(yosay(
     chalk.red('Welcome!') + '\n' +
-    chalk.yellow('You\'re using OpenVeo Plugin Generator !')
+    chalk.yellow('You\'re using OpenVeo Plugin Generator!')
   ));
 
   // Define all questions asked to the user
@@ -37,8 +37,8 @@ module.exports = function() {
     prompts.push({
       type: 'input',
       name: 'plugin',
-      message: 'What name do you want to give to your OpenVeo plugin ?',
-      validate: function(answer) {
+      message: 'What name do you want to give to your OpenVeo plugin?',
+      validate: (answer) => {
         if (!isValidPluginName(answer))
           return 'Plugin name should not use special characters except dashes and should not start by "openveo".';
         else
@@ -47,22 +47,22 @@ module.exports = function() {
     });
   }
 
-  return this.prompt(prompts).then(function(answers) {
-    self.options.pluginName = answers.plugin || self.options.pluginName;
+  const answers = await this.prompt(prompts);
+  this.options.pluginName = answers.plugin || this.options.pluginName;
 
-    if (!self.options.pluginName || !isValidPluginName(self.options.pluginName))
-      this.env.error(chalk.red('Invalid plugin name !'));
+  if (!this.options.pluginName || !isValidPluginName(this.options.pluginName))
+    this.env.error(chalk.red('Invalid plugin name!'));
 
-    this.properties.answers = answers;
+  this.properties.answers = answers;
 
-    this.properties.templated.pluginName = 'openveo-' + self.options.pluginName;
-    this.properties.templated.originalPluginName = self.options.pluginName;
-    this.properties.templated.plugin = _.camelCase(_.lowerCase(self.options.pluginName));
-    this.properties.templated.Plugin = _.upperFirst(_.camelCase(self.options.pluginName));
-    this.properties.templated.PLUGIN = _.toUpper(_.snakeCase(self.options.pluginName));
+  this.properties.templated.pluginName = `openveo-${this.options.pluginName}`;
+  this.properties.templated.originalPluginName = this.options.pluginName;
+  this.properties.templated.plugin = _.camelCase(_.lowerCase(this.options.pluginName));
+  this.properties.templated.Plugin = _.upperFirst(_.camelCase(this.options.pluginName));
+  this.properties.templated.PLUGIN = _.toUpper(_.snakeCase(this.options.pluginName));
 
-    this.properties.templated.project = paths.project;
-    this.properties.templated.pkg = paths.pkg;
+  this.properties.templated.project = paths.project;
+  this.properties.templated.pkg = paths.pkg;
 
-  }.bind(this));
+  return Promise.resolve();
 };