+++ generators/app/src/writing.js
@@ -1,55 +1,54 @@
 'use strict';
 
-var path = require('path');
+const path = require('path');
 
-var config = require('../config/files');
+const config = require('../config/files');
 
-var UNDERSCORE = '_';
-var PATTERN_LOWERCASE = /\{\{plugin\}\}/;
-var PATTERN_CAPITALIZE = /\{\{Plugin\}\}/;
-var PATTERN_ORIGINAL = /\{\{originalPluginName\}\}/;
+const UNDERSCORE = '_';
+const PATTERN_LOWERCASE = /\{\{plugin\}\}/;
+const PATTERN_CAPITALIZE = /\{\{Plugin\}\}/;
+const PATTERN_ORIGINAL = /\{\{originalPluginName\}\}/;
 
-module.exports = function() {
+module.exports = async function() {
+  const pluginLowercase = this.properties.templated.plugin;
+  const pluginCapitalize = this.properties.templated.Plugin;
+  const pluginOriginal = this.properties.templated.originalPluginName;
 
-  var pluginLowercase = this.properties.templated.plugin;
-  var pluginCapitalize = this.properties.templated.Plugin;
-  var pluginOriginal = this.properties.templated.originalPluginName;
-
-  var plugin = function(dest) {
+  const plugin = (dest) => {
     dest = dest.replace(PATTERN_LOWERCASE, pluginLowercase);
     dest = dest.replace(PATTERN_CAPITALIZE, pluginCapitalize);
     dest = dest.replace(PATTERN_ORIGINAL, pluginOriginal);
-
     return dest;
   };
 
-  var copyFile = function(source, destination) {
+  const copyFile = (source, destination) => {
     destination = ('string' === typeof destination) ? destination : source;
-    this.fs.copy(
+    return this.fs.copyAsync(
       this.templatePath(source),
-      plugin(path.join(this.properties.url, destination))
+      plugin(this.destinationPath(destination))
     );
-  }.bind(this);
-
-  var copyTemplatedFile = function(file) {
-    var basename = path.basename(file);
-    var prefixed = file.replace(basename, UNDERSCORE + basename);
-
-    this.fs.copyTpl(
-      this.templatePath(prefixed),
-      plugin(path.join(this.properties.url + file)),
-      this.properties.templated
-    );
-  }.bind(this);
-
-  var copyDotFile = function(file) {
-    var basename = path.basename(file);
-    var prefixed = file.replace(basename, basename.substring(1));
-    copyFile(prefixed, file);
   };
 
-  config.src.forEach(copyFile);
-  config.tasks.forEach(copyFile);
-  config.templated.forEach(copyTemplatedFile);
-  config.dots.forEach(copyDotFile);
+  const copyTemplatedFile = (file) => {
+    const basename = path.basename(file);
+    const prefixed = file.replace(basename, UNDERSCORE + basename);
+
+    return this.fs.copyTplAsync(
+      this.templatePath(prefixed),
+      plugin(this.destinationPath(file)),
+      this.properties.templated
+    );
+  };
+
+  const copyDotFile = (file) => {
+    const basename = path.basename(file);
+    const prefixed = file.replace(basename, basename.substring(1));
+    return copyFile(prefixed, file);
+  };
+
+  await Promise.all(config.src.map((src) => copyFile(src)));
+  await Promise.all(config.tasks.map((task) => copyFile(task)));
+  await Promise.all(config.templated.map((templatedFile) => copyTemplatedFile(templatedFile)));
+  await Promise.all(config.dots.map((dotFile) => copyDotFile(dotFile)));
+  return Promise.resolve();
 };