+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - **grunt-eslint** has been removed
 - **grunt-gh-pages** has been removed
 - **grunt-mkdocs** has been removed
+- **yeoman-generator** has been upgraded from 1.0.1 to **5.4.2**
 
 # 5.0.0 / 2020-05-04